change double and add test

+++ src/main/java/com/tset/calculator/dto/CalculationRequest.java
@@ -1,13 +1,13 @@
 package com.tset.calculator.dto;
 
 public record CalculationRequest(
-        float principal,
-        float annualInterestRate,
+        double principal,
+        double annualInterestRate,
         int timesPerYear,
         int years,
-        float monthlyContribution,
-        float monthlyContributionIncreaseAnnually,
-        float housePrice,
-        float annualHousePriceIncrease,
-        float mortgageAmount
+        double monthlyContribution,
+        double monthlyContributionIncreaseAnnually,
+        double housePrice,
+        double annualHousePriceIncrease,
+        double mortgageAmount
 ) {}
